ref(seer): move DatadogPatConnectModal into sentry/ for reuse
Relocate the Datadog PAT connect modal from getsentry into open-source
sentry/components/seer/ so it can be reused outside the billing-gated
Seer Automation settings page (e.g. the Seer Explorer reauth flow).

Pure relocation: the modal depends only on sentry/* and @sentry/scraps/*,
so it works anywhere the open-source monitoring-provider endpoints do.
The settings section (MonitoringProvidersSection) stays in getsentry and
now imports the modal from its new location.
